change IdType to ASSIGN_ID(SnowFlake)

+++ mybatisplus/src/main/java/com/lim/test/mybatisplus/role/controller/RoleController.java
@@ -53,7 +53,7 @@ public class RoleController {
      * @return 影响的记录条数
      */
     @DeleteMapping
-    public Integer del(Integer id) {
+    public Integer del(Long id) {
         return roleService.delete(id);
     }
