+++ mybatisplus/src/main/java/com/lim/test/mybatisplus/role/service/impl/RoleServiceImpl.java
@@ -42,7 +42,7 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements IR
     }
 
     @Override
-    public Integer delete(Integer id) {
+    public Integer delete(Long id) {
         return baseMapper.deleteById(id);
     }
 }
