+++ mybatisplus/src/main/java/com/lim/test/mybatisplus/user/controller/UserController.java
@@ -106,7 +106,7 @@ public class UserController {
      * @return 影响的记录条数
      */
     @DeleteMapping
-    public Integer del(Integer id) {
+    public Integer del(Long id) {
         return userService.delete(id);
     }
